+++ examples/platformer_2d.rs
@@ -228,7 +228,7 @@ fn setup_player(mut commands: Commands) {
             coyote_time: 0.15,
             jump_input_buffer_time: 0.2,
             held_jump_cooldown: None,
-            jump_start_extra_gravity: 30.0,
+            upslope_jump_extra_gravity: 30.0,
             jump_takeoff_extra_gravity: 30.0,
             jump_takeoff_above_velocity: 3.0,
             jump_fall_extra_gravity: 20.0,